Add files via upload
The primary objective of the CO PO Attainment Software is to facilitate the systematic assessment and measurement of Course Outcomes (COs) and Program Outcomes (POs) in an educational institution. This software aims to streamline the process of data entry, evaluation, and reporting to ensure continuous improvement in the quality of education.